+++ EpikCord/client/__init__.py
@@ -1,9 +1,8 @@
-from .client import Client, TokenStore
+from .client import Client
 from .http import APIVersion, HTTPClient
 
 __all__ = (
     "APIVersion",
     "Client",
-    "TokenStore",
     "HTTPClient",
 )
